% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTopGameIDs.R
\name{getTopGameIDs}
\alias{getTopGameIDs}
\title{Twitch ETL: Obtain top games}
\usage{
getTopGameIDs(clientID, numGames = 100)
}
\arguments{
\item{clientID}{Twitch API client key (str)}

\item{numGames}{}
}
\value{
A dataframe containing the game IDs, game names, and pull date
}
\description{
Twitch ETL: Obtain top games
}
\examples{
getTopGameIDs('api_key', 10)
}
